Project2 Exercise - Small fix
The following is the exercise from the book:
"Given two integers "start" and "end," where "end" is greater than "start," write a recursive Java method that returns the sum of the integers from "start" through "end," inclusive."
The attached code adds the integers between two integers ONLY if the smaller integer is entered first. For example, if the user inputs 2 and then 5, the result that is displayed is 14, since 2 + 3 + 4 + 5 = 14. What I tried to do in the next "else if" is to switch the inputs making temp1 the smaller integer and then call the method again with the new modified parameters. The problem is, every time I run it, the program crashes. Please take a look at it.